Move async_lru_cache and improve network.py
